+++ kicost/edas/tools.py
@@ -28,8 +28,7 @@ __company__ = 'University of Campinas - Brazil'
 
 # Libraries.
 import re, os # Regular expression parser and matches.
-from ..global_vars import logger, DEBUG_OVERVIEW, DEBUG_DETAILED, DEBUG_OBSESSIVE # Debug configurations.
-from ..global_vars import SEPRTR
+from ..global_vars import * # Debug, language and default configurations.
 from ..distributors.global_vars import distributor_dict
 from . import eda_dict # EDA dictionary with the features.