+++ HISTORY.rst
@@ -8,6 +8,7 @@ ______________________
 * Fix minimum order number and warning the user, it also show in the price break.
 * Improve some spreadsheet formulas.
 * Fix the Kispace query API to prioritize the catalogue number under the manufacture code.
+* Added a global field for purchase information, it will be added as description to be printed into the package labels (Digikey and others distributors that allow to include description into the package labels).
 
 
 1.1 (2019-07-15)